+++ lib/character/CharacterPainter.dart
