+++ lib/character/CharacterState.dart
